don't store frame's location to layout-local
This would normally not be a problem, except the game will anchor a dragged frame to the nearest point of the parent frame, instead of just changing the offset on existing points. That can change the behavior of SetHeight (i.e. the frame grows/shrinks relative to the anchor point, which is fixed in place), and I'd like for the maximinimize button to never move the title bar. Thus, we need the anchor point to always be a TOP & can't rely on the builtin layout cache.